+++ crates/parry2d/examples/proximity_query2d.rs
@@ -10,6 +10,14 @@ fn main() {
     let ball_pos_intersecting = Pose::translation(1.0, 1.0);
     let ball_pos_disjoint = Pose::translation(3.0, 3.0);
 
-    assert!(query::intersection_test(&ball_pos_intersecting, &ball, &cuboid_pos, &cuboid).unwrap());
-    assert!(!query::intersection_test(&ball_pos_disjoint, &ball, &cuboid_pos, &cuboid).unwrap());
+    assert!(
+        query::intersection_test(&ball_pos_intersecting, &ball, &cuboid_pos, &cuboid)
+            .unwrap()
+            .intersecting
+    );
+    assert!(
+        !query::intersection_test(&ball_pos_disjoint, &ball, &cuboid_pos, &cuboid)
+            .unwrap()
+            .intersecting
+    );
 }